agentforce-voice-main/requirements.txt
Flask==3.0.0
requests==2.32.5
python-dotenv==1.1.1
gunicorn==21.2.0
Flask-Cors==4.0.0